requirements.txt
yolk
cffi
greenlet
numpy==1.9.0
pyparsing
pytest-xdist
